v.1.6.7.

+++ HybPhyloMaker0c_Rsetup_MetaCentrum.sh
@@ -10,19 +10,25 @@
 # *    HybPhyloMaker - Pipeline for Hyb-Seq data processing and tree building    *
 # *                  https://github.com/tomas-fer/HybPhyloMaker                  *
 # *                 Script 0c - Setup R packages on Metacentrum                  *
-# *                                   v.1.6.6                                    *
-# * Tomas Fer, Dept. of Botany, Charles University, Prague, Czech Republic, 2018 *
+# *                                   v.1.6.7                                    *
+# * Tomas Fer, Dept. of Botany, Charles University, Prague, Czech Republic, 2021 *
 # * tomas.fer@natur.cuni.cz                                                      *
 # ********************************************************************************
 
 #Download working versions of specific packages and install them to personal library
-#ape 3.5 (IMPORTANT, v 3.4 is not compatible with HybPhyloMaker)
+#ape 5.0 (IMPORTANT, v 3.4 is not compatible with HybPhyloMaker)
 #data.table 1.9.6
 #phytools 0.5-20
 #seqinr 3.1-3
 #openxlsx 4.0.0
+#phangorn 2.5.5 (this works also with R 3.5 and lower)
 
-echo -e "\nScript HybPhyloMaker0c is running on Metacentrum..."
+if [[ $PBS_O_HOST == *".cz" ]]; then
+	echo -e "\nScript HybPhyloMaker0c is running on Metacentrum..."
+else
+	echo -e "\nYou are not on Metacentrum. Install the R packages using 'install_software.sh' script. Exiting...\n"
+	exit 3
+fi
 
 #Move to scratch
 cd $SCRATCHDIR
@@ -40,7 +46,9 @@ wget https://cran.r-project.org/src/contrib/Archive/data.table/data.table_1.9.6.
 wget https://cran.r-project.org/src/contrib/Archive/phytools/phytools_0.5-20.tar.gz
 wget https://cran.r-project.org/src/contrib/Archive/seqinr/seqinr_3.1-3.tar.gz
 wget https://cran.r-project.org/src/contrib/Archive/openxlsx/openxlsx_4.0.0.tar.gz
-wget https://cran.r-project.org/src/contrib/Archive/phangorn/phangorn_2.5.5.tar.gz
+wget https://cran.r-project.org/src/contrib/Archive/quadprog/quadprog_1.5-7.tar.gz #required by phangorn
+wget https://cran.r-project.org/src/contrib/Archive/igraph/igraph_1.2.5.tar.gz #required by phangorn
+wget https://cran.r-project.org/src/contrib/fastmatch_1.1-0.tar.gz #required by phangorn
 
 #Install packages to personal (writable) library
 mkdir -p /storage/$server/home/$LOGNAME/Rpackages
@@ -49,10 +57,18 @@ for package in $(ls *.tar.gz); do
 	R CMD INSTALL $package -l /storage/$server/home/$LOGNAME/Rpackages
 done
 
+wget https://cran.r-project.org/src/contrib/Archive/phangorn/phangorn_2.5.5.tar.gz
+R CMD INSTALL phangorn_2.5.5.tar.gz -l /storage/$server/home/$LOGNAME/Rpackages #this only works once the previous packeges are installed
+
 #Use this command before running R in HybPhyloMaker scripts
 export R_LIBS="/storage/$server/home/$LOGNAME/Rpackages"
 
 #Delete scratch
-rm -rf $SCRATCHDIR/*
+if [[ $PBS_O_HOST == *".cz" ]]; then
+	#delete scratch
+	if [[ ! $SCRATCHDIR == "" ]]; then
+		rm -rf $SCRATCHDIR/*
+	fi
+fi
 
 echo -e "\nScript HybPhyloMaker0c finished...\n"
